+++ backend/src/main/java/uit/ac/ma/est/kessabpro/repositories/TransactionRepository.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 public interface TransactionRepository extends JpaRepository<Transaction, UUID> {
 
     @Query("SELECT t.amount FROM Transaction t WHERE t.sale.id = :saleId")
-    List<BigDecimal> findAmountsBySaleId(UUID saleId);
+    List<Double> findAmountsBySaleId(UUID saleId);
 
     List<Transaction> getTransactionBySaleId(UUID saleId);
 }
